Fix Transaction ViewModel Locations
  - Moved Transaction ViewModels and AutoMapper Profiles from BLL to Web Layer
