+++ AltFuture.BusinessLogicLayer/AutoMapper/MarketDataPriceToCryptoPriceProfile.cs
@@ -1,21 +1,16 @@
 ﻿using AltFuture.DataAccessLayer.Models;
 using AltFuture.MarketDataConsumer.Models.MarketDataClient;
 using AutoMapper;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace AltFuture.BusinessLogicLayer.AutoMapper
+
+namespace AltFuture.BusinessLogicLayer.AutoMapper;
+
+public class MarketDataPriceToCryptoPriceProfile : Profile
 {
-    public class MarketDataPriceToCryptoPriceProfile : Profile
+    public MarketDataPriceToCryptoPriceProfile()
     {
-        public MarketDataPriceToCryptoPriceProfile()
-        {
-            CreateMap<MarketPriceData,CryptoPrice>()
-                .ForMember(dest => dest.Price, opt => opt.MapFrom(src => src.MarketPrice))
-                ;
-        }
+        CreateMap<MarketPriceData,CryptoPrice>()
+            .ForMember(dest => dest.Price, opt => opt.MapFrom(src => src.MarketPrice))
+            ;
     }
 }
