+++ AltFuture.WebApp/Areas/Portfolios/AutoMapper/TransactionTransactionCreateProfile.cs
@@ -1,0 +1,16 @@
+﻿using AutoMapper;
+using AltFuture.DataAccessLayer.Models;
+using AltFuture.WebApp.Areas.Portfolios.ViewModels;
+
+namespace AltFuture.WebApp.Areas.Portfolios.AutoMapper;
+
+public class TransactionTransactionCreateProfile: Profile
+{
+
+    public TransactionTransactionCreateProfile()
+    {
+        CreateMap<Transaction, TransactionCreate>()
+            .ReverseMap();                
+    }
+
+}
