+++ AltFuture.WebApp/Areas/Portfolios/Views/Transactions/Create.cshtml
@@ -1,5 +1,4 @@
-﻿@using AltFuture.BusinessLogicLayer.ViewModels.Transactions
-@model TransactionCreate
+﻿@model TransactionCreate
 
 @{
     ViewData["Title"] = "Create Transaction";
