+++ AltFuture.WebApp/Areas/Portfolios/Views/Transactions/Details.cshtml
@@ -1,5 +1,4 @@
-﻿@using AltFuture.BusinessLogicLayer.ViewModels.Transactions
-@model TransactionDetail
+﻿@model TransactionDetail
 
 @{
     ViewData["Title"] = "Transaction Details";
