+++ AltFuture.WebApp/Areas/Portfolios/Views/Transactions/Edit.cshtml
@@ -1,5 +1,4 @@
-﻿@using AltFuture.BusinessLogicLayer.ViewModels.Transactions
-@model TransactionEdit
+﻿@model TransactionEdit
 
 @{
     ViewData["Title"] = "Edit Transaction";
