+++ AltFuture.WebApp/Areas/Portfolios/Views/_ViewImports.cshtml
@@ -1,4 +1,5 @@
 ﻿@using AltFuture.WebApp
 @using AltFuture.WebApp.ViewModels
+@using AltFuture.WebApp.Areas.Portfolios.ViewModels
 @using AltFuture.DataAccessLayer.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
